worklog: cleanup: use a single variable to list boards
Instead of listing the possible boards all over the code, cosolidate
most of these lists into a single variable that contains all active
kanban boards. In theory, this variable could be changed to add
additional boards; this is not currently possible without changes to the
dashboard view, but could be added in the future.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>