worklog: tags: filter the dashboard by tag
Add a new variable for filtering the dashboard. If the tag filter is
non-nil, then refreshing the dashboard will only show worklogs that have
a tag matching the filter. The filter applies globally and will be used
on all subsequent refreshes of the dashboard.

To set the filter, use the worklog-dashboard-for-tag command, which will
set the filter to the given tag, then (re-)open the dashboard with the
new filter. Set the filter to the empty string in order to clear the
filter and start displaying all worklogs again.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>