urxvt: scale font size with dpi
Get the dpi resolution of the display using xdpyinfo and scale the font
size for urxvt based on dpi instead of using a fixed size. I suspect
that the scaling will need to change a bit in the future, but (integer)
division by 6.5 seems to give adequate results on both desktop and
laptop.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>