i3status: initial revision
Install and configure i3status for i3bar. The current setup generates
the 'skydio uptime' counter in guile and shells out to i3status for the
other data. Parts of this will need to change for other computers
(specifically, the networking interface).

To enable this, use the following in ~/.i3/config:

bar {
    status_command guile -s ~/.i3/status.scm
    ...
}

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>