nix: install system-wide
Using nix from within guix requires a system-wide service for managing
the nix store. This means that nix should be a system-level package.

I'm not sure this is strictly required, since you may be able to create
a user-level nix service; I'm unsure if that will work correctly though,
and the system-level daemon seems a little cleaner for redtail.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>