home: i3: take the status script as a config parameter
If we want to use home/ as a load-path for the 'guix home' invocation,
we need to remove scheme files that shouldn't be loaded, including
i3-status.scm. In order to do this, the actual top-level home
environment definition needs to define where to find this
script. Making it a config parameter is the simplest way to do this.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>