home: i3: add keybinding for i3lock
The actual i3lock package is not managed through the home
environment. Since the screen locker must be setuid-root and needs a
pam service, it must be installed system-wide, either via the
screen-locker service or the host's package manager.

On work machines, this means relying on the host i3lock package.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>