home: i3: allow setting browser bin name
The main binary for a package will not necessarily match the package
name, so always using the package name in my i3 config is
incorrect. Add an optional config value for setting the web browser
binary name. This is used on my skydio config for binary name
"chromium" in the "ungoogled-chromium" package.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>