packages: revup: Initial revision
Add an initial revision of the revup package v0.1.1. This requires
patching a handful of components to be compatible with newer versions
of setuptools and remove some dependencies that are included in core
python.

I've merged these patches upstream; once there is a new release, only
the setup.py patch and the substitutions will be required. I'm not
merging the setup.py upstream, since using a setup.py directly is
discouraged (even though guix still uses it).

Signed-off-by: Brian Kubisiak <brian.kubisiak@skydio.com>