home: git: Fix 'Fixes:' trailers
I misunderstood how ifMissing and ifExists are implemented: the
command is first run without an argument to add the missing trailer,
then re-run with an argument to fix it. I need to replace the existing
trailer in order to get what I want.

This unfortunately means that I can't have multiple 'Fixes:' trailers,
but that should be fine.

Fixes: commit a095216 ("home: git: Add 'Fixes:' trailer section")
Signed-off-by: Brian Kubisiak <brian@kubisiak.com>