home: rsnapshot: Configure backup levels
Instead of hardcoding these, allow the user to configure these. This
centralizes the source-of-truth for both the cron jobs and the
'retain' settings in the config file, making it easier to modify the
jobs that run.

Additionally, add back the beta and gamma jobs to run weekly and
monthly backups, respectively.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>